opt: Remove relational operators from older code.
We don't use relational operators in our index constraints code, so
remove them. This will make another PR changing how the catalog
works easier to land. In the near future, the rest of the older
index constraints code will be removed.

Release note: None
