+++ pkg/sql/opt/build.go
@@ -18,9 +18,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/optbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
 )
 
 // Map from tree.ComparisonOperator to operator.
@@ -145,39 +143,18 @@ var unaryOpReverseMap = [...]tree.UnaryOperator{
 const exprAllocChunk = 16
 const scalarPropsAllocChunk = 16
 
-// TODO(rytaft): Increase the relationalProps chunk size after more relational
-// operators are implemented.
-const relationalPropsAllocChunk = 1
-
 type buildContext struct {
-	ctx     context.Context
-	evalCtx *tree.EvalContext
-	catalog optbase.Catalog
-
-	scope *scope
-
-	preallocScalarProps     []scalarProps
-	preallocExprs           []Expr
-	preallocRelationalProps []relationalProps
+	ctx                 context.Context
+	evalCtx             *tree.EvalContext
+	preallocScalarProps []scalarProps
+	preallocExprs       []Expr
 }
 
-func makeBuildContext(
-	ctx context.Context, evalCtx *tree.EvalContext, catalog optbase.Catalog,
-) buildContext {
+func makeBuildContext(ctx context.Context, evalCtx *tree.EvalContext) buildContext {
 	return buildContext{
 		ctx:     ctx,
 		evalCtx: evalCtx,
-		catalog: catalog,
 	}
-}
-
-func (bc *buildContext) newRelationalProps() *relationalProps {
-	if len(bc.preallocRelationalProps) == 0 {
-		bc.preallocRelationalProps = make([]relationalProps, relationalPropsAllocChunk)
-	}
-	p := &bc.preallocRelationalProps[0]
-	bc.preallocRelationalProps = bc.preallocRelationalProps[1:]
-	return p
 }
 
 func (bc *buildContext) newScalarProps() *scalarProps {
@@ -204,188 +181,6 @@ func (bc *buildContext) newScalarExpr() *Expr {
 	e := bc.newExpr()
 	e.scalarProps = bc.newScalarProps()
 	return e
-}
-
-// newRelationalExpr returns a new *Expr with a new, blank relationalProps.
-func (bc *buildContext) newRelationalExpr() *Expr {
-	e := bc.newExpr()
-	e.relProps = bc.newRelationalProps()
-	return e
-}
-
-// resolve converts expr to a TypedExpr and normalizes the resulting expression.
-// It panics if there are any errors during conversion or normalization.
-func (bc *buildContext) resolve(expr tree.Expr, desired types.T) tree.TypedExpr {
-	texpr := bc.scope.resolve(expr, desired)
-	nexpr, err := bc.evalCtx.NormalizeExpr(texpr)
-	if err != nil {
-		panic(builderError{err})
-	}
-
-	return nexpr
-}
-
-// build converts a tree.Statement to an Expr tree.
-func (bc *buildContext) build(stmt tree.Statement) *Expr {
-	var result *Expr
-	switch stmt := stmt.(type) {
-	case *tree.ParenSelect:
-		result = bc.buildSelect(stmt.Select)
-
-	case *tree.Select:
-		result = bc.buildSelect(stmt)
-
-	default:
-		panic(errorf("unexpected statement: %T", stmt))
-	}
-
-	return result
-}
-
-// buildSelect converts a tree.Select to an Expr tree. This method will
-// expand significantly once we implement joins, aggregations, filters, etc.
-func (bc *buildContext) buildSelect(stmt *tree.Select) *Expr {
-	var result *Expr
-	switch t := stmt.Select.(type) {
-	case *tree.ParenSelect:
-		result = bc.buildSelect(t.Select)
-
-	case *tree.SelectClause:
-		if (t.GroupBy != nil && len(t.GroupBy) > 0) || len(t.Exprs) > 1 || t.Distinct {
-			panic(errorf("complex queries not yet supported."))
-		}
-		result = bc.buildFrom(t.From, t.Where)
-
-	default:
-		panic(errorf("unexpected select statement: %T", stmt.Select))
-	}
-
-	return result
-}
-
-// buildSelect converts a tree.From to an Expr tree. This method
-// will expand significantly once we implement joins.
-func (bc *buildContext) buildFrom(from *tree.From, where *tree.Where) *Expr {
-	if len(from.Tables) != 1 {
-		panic(errorf("joins not yet supported."))
-	}
-	result := bc.buildTable(from.Tables[0])
-	bc.scope = bc.scope.push()
-	bc.scope.cols = append(bc.scope.cols, result.relProps.columns...)
-
-	if where != nil {
-		input := result
-		result = bc.newRelationalExpr()
-		initSelectExpr(result, input)
-		texpr := bc.resolve(where.Expr, types.Bool)
-		result.addFilter(bc.buildScalar(texpr))
-		result.initProps()
-		bc.scope = bc.scope.push()
-		bc.scope.cols = append(bc.scope.cols, result.relProps.columns...)
-	}
-
-	return result
-}
-
-// buildTable converts a tree.TableExpr to an Expr tree.  For example,
-// if the tree.TableExpr consists of a single table, the resulting Expr
-// tree would consist of a single expression with a scanOp operator.
-func (bc *buildContext) buildTable(texpr tree.TableExpr) *Expr {
-	// NB: The case statements are sorted lexicographically.
-	switch source := texpr.(type) {
-	case *tree.AliasedTableExpr:
-		result := bc.buildTable(source.Expr)
-		if source.As.Alias != "" {
-			if n := len(source.As.Cols); n > 0 && n != len(result.relProps.columns) {
-				panic(errorf("rename specified %d columns, but table contains %d",
-					n, len(result.relProps.columns)))
-			}
-
-			for i := range result.relProps.columns {
-				col := &result.relProps.columns[i]
-				if i < len(source.As.Cols) {
-					col.name = columnName(source.As.Cols[i])
-				}
-				col.table = tableName(source.As.Alias)
-			}
-		}
-		return result
-
-	case *tree.FuncExpr:
-		panic(errorf("unimplemented table expr: %T", texpr))
-
-	case *tree.JoinTableExpr:
-		panic(errorf("unimplemented table expr: %T", texpr))
-
-	case *tree.NormalizableTableName:
-		tn, err := source.Normalize()
-		if err != nil {
-			panic(errorf("%s", err))
-		}
-		tab, err := bc.catalog.FindTable(bc.ctx, tn)
-		if err != nil {
-			panic(errorf("%s", err))
-		}
-
-		return bc.buildScan(tab)
-
-	case *tree.ParenTableExpr:
-		return bc.buildTable(source.Expr)
-
-	case *tree.StatementSource:
-		panic(errorf("unimplemented table expr: %T", texpr))
-
-	case *tree.Subquery:
-		return bc.build(source.Select)
-
-	case *tree.TableRef:
-		panic(errorf("unimplemented table expr: %T", texpr))
-
-	default:
-		panic(errorf("unexpected table expr: %T", texpr))
-	}
-}
-
-// buildScan creates an Expr with a scanOp operator for the given table.
-func (bc *buildContext) buildScan(tab optbase.Table) *Expr {
-	result := bc.newRelationalExpr()
-	initScanExpr(result, tab)
-	result.relProps.columns = make([]columnProps, 0, len(tab.TabName()))
-	props := result.relProps
-
-	// Every reference to a table in the query gets a new set of output column
-	// indexes. Consider the query:
-	//
-	//   SELECT * FROM a AS l JOIN a AS r ON (l.x = r.y)
-	//
-	// In this query, `l.x` is not equivalent to `r.x` and `l.y` is not
-	// equivalent to `r.y`. In order to achieve this, we need to give these
-	// columns different indexes.
-	state := bc.scope.state
-	tabName := tableName(tab.TabName())
-	state.tables[tabName] = append(state.tables[tabName], state.nextColumnIndex)
-
-	for i := 0; i < tab.NumColumns(); i++ {
-		col := tab.Column(i)
-		colProps := columnProps{
-			index: state.nextColumnIndex,
-			name:  columnName(col.ColName()),
-			table: tabName,
-			typ:   col.DatumType(),
-		}
-		props.columns = append(props.columns, colProps)
-		state.nextColumnIndex++
-	}
-
-	// Initialize not-NULL columns from the table schema.
-	for i := 0; i < tab.NumColumns(); i++ {
-		if !tab.Column(i).IsNullable() {
-			props.notNullCols.Add(props.columns[i].index)
-		}
-	}
-
-	result.initProps()
-	return result
 }
 
 // buildScalar converts a tree.TypedExpr to an Expr tree.
@@ -497,27 +292,6 @@ func errorf(format string, a ...interface{}) builderError {
 	return builderError{err}
 }
 
-// build converts a tree.Statement to an Expr tree.
-func build(
-	ctx context.Context, stmt tree.Statement, catalog optbase.Catalog, evalCtx *tree.EvalContext,
-) (_ *Expr, err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			if bldErr, ok := r.(builderError); ok {
-				err = bldErr
-			} else {
-				panic(r)
-			}
-		}
-	}()
-	buildCtx := makeBuildContext(ctx, evalCtx, catalog)
-	buildCtx.scope = &scope{
-		state: &queryState{tables: make(map[tableName][]columnIndex)},
-	}
-	buildCtx.scope.state.semaCtx.Placeholders = tree.MakePlaceholderInfo()
-	return buildCtx.build(stmt), nil
-}
-
 // buildScalar converts a tree.TypedExpr to an Expr tree.
 func buildScalar(pexpr tree.TypedExpr, evalCtx *tree.EvalContext) (_ *Expr, err error) {
 	// We use panics in buildScalar code because it makes the code less tedious;
@@ -531,7 +305,7 @@ func buildScalar(pexpr tree.TypedExpr, evalCtx *tree.EvalContext) (_ *Expr, err 
 			}
 		}
 	}()
-	buildCtx := makeBuildContext(context.TODO(), evalCtx, nil /* catalog */)
+	buildCtx := makeBuildContext(context.TODO(), evalCtx)
 	return buildCtx.buildScalar(pexpr), nil
 }
